+++ CHANGELOG.md
@@ -88,6 +88,12 @@
 * Add `EveOnline::ESI::CharacterClones`
 * Add `EveOnline::ESI::Models::JumpClone`
 * Add `EveOnline::ESI::Models::HomeLocation`
+* Drop `EveOnline::JumpClone`
+* Drop `EveOnline::JumpCloneImplant`
+* Drop `EveOnline::CharacterJumpClones`
+* Drop `EveOnline::CharacterJumpCloneImplants`
+* Drop `EveOnline::XML::CharacterSheet#jump_clones`
+* Drop `EveOnline::XML::CharacterSheet#jump_clone_implants`
 
 **v0.12.0**